requirements.txt
fastapi
uvicorn
pillow
numpy
ultralytics
torch
torchvision
easyocr
python-dotenv
